src: introduce NodeOption
Current command line options are stored using global state.
This introduces a new internal `NodeOption` class that is
associated with the `Environment` to store startup options.

Note: "associated with the `Environment` does not mean owned
by the `Environment` or per-Environment.

This is the first step in refactoring options parsing and
process bootstrap.
