+++ node.gyp
@@ -309,6 +309,7 @@
         'src/node_file.cc',
         'src/node_http2.cc',
         'src/node_http_parser.cc',
+        'src/node_options.cc',
         'src/node_os.cc',
         'src/node_platform.cc',
         'src/node_perf.cc',
@@ -365,6 +366,7 @@
         'src/node_internals.h',
         'src/node_javascript.h',
         'src/node_mutex.h',
+        'src/node_options.h',
         'src/node_platform.h',
         'src/node_perf.h',
         'src/node_perf_common.h',
@@ -984,6 +986,7 @@
         '<(obj_path)<(obj_separator)handle_wrap.<(obj_suffix)',
         '<(obj_path)<(obj_separator)env.<(obj_suffix)',
         '<(obj_path)<(obj_separator)node.<(obj_suffix)',
+        '<(obj_path)<(obj_separator)node_options.<(obj_suffix)',
         '<(obj_path)<(obj_separator)node_buffer.<(obj_suffix)',
         '<(obj_path)<(obj_separator)node_debug_options.<(obj_suffix)',
         '<(obj_path)<(obj_separator)node_i18n.<(obj_suffix)',
