+++ src/node_buffer.h
@@ -27,8 +27,6 @@
 
 namespace node {
 
-extern bool zero_fill_all_buffers;
-
 namespace Buffer {
 
 static const unsigned int kMaxLength =
