+++ src/node_constants.h
@@ -41,34 +41,9 @@
 #define RSA_PSS_SALTLEN_AUTO -2
 #endif
 
-#define DEFAULT_CIPHER_LIST_CORE "ECDHE-RSA-AES128-GCM-SHA256:"     \
-                                 "ECDHE-ECDSA-AES128-GCM-SHA256:"   \
-                                 "ECDHE-RSA-AES256-GCM-SHA384:"     \
-                                 "ECDHE-ECDSA-AES256-GCM-SHA384:"   \
-                                 "DHE-RSA-AES128-GCM-SHA256:"       \
-                                 "ECDHE-RSA-AES128-SHA256:"         \
-                                 "DHE-RSA-AES128-SHA256:"           \
-                                 "ECDHE-RSA-AES256-SHA384:"         \
-                                 "DHE-RSA-AES256-SHA384:"           \
-                                 "ECDHE-RSA-AES256-SHA256:"         \
-                                 "DHE-RSA-AES256-SHA256:"           \
-                                 "HIGH:"                            \
-                                 "!aNULL:"                          \
-                                 "!eNULL:"                          \
-                                 "!EXPORT:"                         \
-                                 "!DES:"                            \
-                                 "!RC4:"                            \
-                                 "!MD5:"                            \
-                                 "!PSK:"                            \
-                                 "!SRP:"                            \
-                                 "!CAMELLIA"
 #endif
 
 namespace node {
-
-#if HAVE_OPENSSL
-extern const char* default_cipher_list;
-#endif
 
 void DefineConstants(v8::Isolate* isolate, v8::Local<v8::Object> target);
 }  // namespace node
