+++ src/node_main.cc
@@ -20,6 +20,7 @@
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #include "node.h"
+#include "node_options.h"
 #include <stdio.h>
 
 #ifdef _WIN32
@@ -33,6 +34,8 @@ int wmain(int argc, wchar_t *wargv[]) {
                     "Windows Server 2008 R2, or higher.");
     exit(ERROR_EXE_MACHINE_TYPE_MISMATCH);
   }
+
+  node::NodeOptions options;
 
   // Convert argv to UTF8
   char** argv = new char*[argc + 1];
@@ -69,7 +72,7 @@ int wmain(int argc, wchar_t *wargv[]) {
   }
   argv[argc] = nullptr;
   // Now that conversion is done, we can finally start.
-  return node::Start(argc, argv);
+  return node::Start(argc, argv, &options);
 }
 #else
 // UNIX
@@ -83,26 +86,24 @@ int wmain(int argc, wchar_t *wargv[]) {
 extern char** environ;
 #endif  // __linux__
 
-namespace node {
-  extern bool linux_at_secure;
-}  // namespace node
-
 int main(int argc, char *argv[]) {
+  int flags = 0;
 #if defined(__linux__)
+  // The AT_SECURE check must occur here, before any further initialization
+  // happens because libuv makes changes that will throw this off.
   char** envp = environ;
   while (*envp++ != nullptr) {}
   Elf_auxv_t* auxv = reinterpret_cast<Elf_auxv_t*>(envp);
   for (; auxv->a_type != AT_NULL; auxv++) {
-    if (auxv->a_type == AT_SECURE) {
-      node::linux_at_secure = auxv->a_un.a_val;
-      break;
-    }
+    if (auxv->a_type == AT_SECURE && auxv->a_un.a_val)
+      flags = node::NODE_OPTION_AT_SECURE;
   }
 #endif
+  node::NodeOptions options(flags);
   // Disable stdio buffering, it interacts poorly with printf()
   // calls elsewhere in the program (e.g., any logging from V8.)
   setvbuf(stdout, nullptr, _IONBF, 0);
   setvbuf(stderr, nullptr, _IONBF, 0);
-  return node::Start(argc, argv);
+  return node::Start(argc, argv, &options);
 }
 #endif
